+++ cloud/objstore/objcli/objgcp/client.go
@@ -267,8 +267,28 @@ func (c *Client) deleteObjects(ctx context.Context, bucket string, objects ...at
 }
 
 func (c *Client) DeleteDirectory(ctx context.Context, opts objcli.DeleteDirectoryOptions) error {
-	fn := func(attrs attrs) error {
-		return c.deleteObjects(ctx, opts.Bucket, attrs)
+	var (
+		// size matches the batch deletion size in AWS/Azure.
+		size  = 1000
+		batch = make([]attrs, 0, size)
+	)
+
+	fn := func(obj attrs) error {
+		batch = append(batch, obj)
+
+		if len(batch) < size {
+			return nil
+		}
+
+		err := c.deleteObjects(ctx, opts.Bucket, batch...)
+		if err != nil {
+			return fmt.Errorf("failed to delete batch: %w", err)
+		}
+
+		clear(batch)
+		batch = batch[:0]
+
+		return nil
 	}
 
 	err := c.iterateObjects(
@@ -281,8 +301,17 @@ func (c *Client) DeleteDirectory(ctx context.Context, opts objcli.DeleteDirector
 		nil,
 		fn,
 	)
+	if err != nil {
+		return fmt.Errorf("failed to iterate objects: %w", err)
+	}
 
-	return err
+	// Ensure we flush the last batch
+	err = c.deleteObjects(ctx, opts.Bucket, batch...)
+	if err != nil {
+		return fmt.Errorf("failed to flush batch: %w", err)
+	}
+
+	return nil
 }
 
 func (c *Client) IterateObjects(ctx context.Context, opts objcli.IterateObjectsOptions) error {